update  StarwarsQueryExecutor failing tests
